Replace `verificationMethods` type in `did:dht` to use `VerificationMethodSpec`.
