+++ dids/src/main/kotlin/web5/sdk/dids/methods/ion/models/Models.kt
@@ -3,13 +3,8 @@ package web5.sdk.dids.methods.ion.models
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
-import com.fasterxml.jackson.annotation.JsonValue
 import com.fasterxml.jackson.core.JsonGenerator
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.DeserializationContext
-import com.fasterxml.jackson.databind.JsonDeserializer
-import com.fasterxml.jackson.databind.JsonSerializer
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
@@ -18,7 +13,8 @@ import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.nimbusds.jose.jwk.JWK
 import web5.sdk.common.Convert
 import web5.sdk.common.EncodingFormat
-import web5.sdk.dids.PublicKeyPurpose
+import web5.sdk.dids.JacksonJwk
+import web5.sdk.dids.PublicKey
 
 /**
  * Represents an ION document containing public keys and services. See bullet 2 in https://identity.foundation/sidetree/spec/#replace.
@@ -44,49 +40,6 @@ public data class Service(
   public val type: String,
   public val serviceEndpoint: String
 )
-
-/**
- * Represents a public key in the ION document as defined in item 3 of https://identity.foundation/sidetree/spec/#add-public-keys
- */
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public data class PublicKey(
-  public val id: String,
-  public val type: String,
-  public val controller: String? = null,
-
-  @JsonSerialize(using = JacksonJwk.Serializer::class)
-  @JsonDeserialize(using = JacksonJwk.Deserializer::class)
-  public val publicKeyJwk: JWK,
-  public val purposes: Iterable<PublicKeyPurpose> = emptyList()
-)
-
-/**
- * JacksonJWK is a utility class that facilitates serialization for [JWK] types, so that it's easy to integrate with any
- * class that is meant to be serialized to/from JSON.
- */
-private class JacksonJwk {
-  /**
-   * [Serializer] implements [JsonSerializer] for use with the [JsonSerialize] annotation from Jackson.
-   */
-  object Serializer : JsonSerializer<JWK>() {
-    override fun serialize(value: JWK, gen: JsonGenerator, serializers: SerializerProvider) {
-      with(gen) {
-        writeObject(value.toJSONObject())
-      }
-    }
-  }
-
-  /**
-   * [Deserializer] implements [JsonDeserializer] for use with the [JsonDeserialize] annotation from Jackson.
-   */
-  object Deserializer : JsonDeserializer<JWK>() {
-    override fun deserialize(p: JsonParser, ctxt: DeserializationContext?): JWK {
-      val typeRef = object : TypeReference<HashMap<String, Any>>() {}
-      val node = p.readValueAs(typeRef) as HashMap<String, Any>
-      return JWK.parse(node)
-    }
-  }
-}
 
 /**
  * Sealed class representing a patch action in the ION document. See https://identity.foundation/sidetree/spec/#did-state-patches
